adjust the installation package GUI so it doesn't install during and installation (ditto update)
closes PR 14263

this may be a release branch candidate. I'm not sure myself, but the PR
sounds worrying.

If it should be considered, then this is the relevant diff (since more than
half of the diff is whitespace changes) to the file pkg/gui/main:

@@ -57,6 +57,8 @@

   (define terminal #f)
   (define (in-terminal-panel abort-label thunk)
+    (cond
+      [(or (not terminal) (send terminal can-close?))
     (send dlg begin-container-sequence)
     (when terminal (send terminal close))
     (define t (in-terminal
@@ -71,7 +73,10 @@
     (disallow-close)
     (send dlg end-container-sequence)
     (yield (send t can-close-evt))
-    (allow-close))
+       (allow-close)]
+      [else
+       (message-box (string-constant install-pkg-dialog-title)
+                    (string-constant install-pkg-not-rentrant))]))
(cherry picked from commit 5059db47d3cd2b42820c12b6cd493da3913b7536)