scribble/manual: make `racketblock` fall back to `racket`-style spacing
When source-location information isn't available, usually due to macros
in compiled modules, make `racketblock` fall back to `racket`-like
spacing. This improvement is visible, for example, in the documentation
for functions like `recycle-icon`.

Using `quote-syntax/srcloc` in a macro is probably a better solution,
because that will preserve the original layout including line breaks,
but the default-spacing fallback seems like an improvement over
unreadable output.