Fix Plot for new undefined behavior: #<undefined> is not a truth value
This is a nice example of why having #<undefined> as a language value is
generally a bad idea. Because in Scheme/Racket, everything that isn't
`refresh?` argument entirely - it was as if its value was always #t. Any
GUI change that was not meant to cause a refresh caused one anyway: a
silent performance error.