places: fix gc of place channels containing place channels
The bug was a kind of typo: using `&` where `%` was intended to
implement a counter wraparound.

This bug is an even more likely candidate to be resopnsible for the
occassional crashes from the DrRacket easter-egg test.
(cherry picked from commit 6d4c25a32241117f0f49d2f071c202906a33bbf2)