make documentation start page tell you when there's a redirecting cookie
If you have anything user-specific installed, the `raco setup` generates
a user-specific documentation page. Visiting that page sets a cookie
(actually, browser local storage, when supported) that redirects searches
and "top" to the user-specific page.

The main installation-specific page now tells you when a redirect is
in place, and it gives you the option of forgetting it (which is handy
if you've removed your user-specific documentation directory, for example).