scribble: use 'pdf-bytes+bounds8, etc.
Use 'pdf-bytes+bounds8, 'png-bytes+bounds8, etc. to convert images
for Latex/PDF and HTML output. The new conversions particularly
help with picts that draw text, since the baseline of the text
can be handled properly in HTML output.

The new conversion also helps hide the difference between a bounding
box and the "inked" area of a pict, since picts normally ink only
slightly outside the bounding box, and the default for pict conversion
pads the image with a few pixels/units in each direction.

Also, add a `++convert <fmt>` command-line option to select a
preferred target for image conversions. (This preference can be
overridden by the existing `render-convertible-as` style property
for HTML rendering.)