raco setup: constrain Scribbled document names
Contrain the fourth element of a list describing a document in
`scribblings` (in "info.rkt") to a collection-path element, instead of
allowing an arbitrary relative path. This constraint is needed by
various parts of the build system, which assume (for example) that all
documents are rendered to an immediate subdirectory.